// generated/src/app/api-models/stage-psclm-dtl-spinal.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StagePsclmDtlSpinal {
  batchId: string;
  transactionId: number;
  lineNo: number;
  subLineCode: string;
  sequenceNo: number;
  spinalManipConditionCode: string;
  spinalManipConditionDesc1: string;
  spinalManipConditionDesc2: string;
  spinalManipConditionXrayAv: string;
  insertDatetime: Date;
}